Moved everything into one place
